Move the editor to it's own crate.
