+++ pad/editor/Cargo.toml
@@ -1,0 +1,53 @@
+[package]
+name = "uiua-editor"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+leptos = { version = "0.6.10", features = ["csr"] }
+js-sys = "0.3.69"
+web-sys = { version = "0.3.70", features = [
+    "Window",
+    "CustomElementRegistry",
+    "CssStyleDeclaration",
+    "DomRect",
+    "Storage",
+    "HtmlAudioElement",
+    "HtmlBrElement",
+    "Selection",
+    "Node",
+    "Clipboard",
+    "ClipboardEvent",
+    "DataTransfer",
+    "Navigator",
+    "Permissions",
+    "ScrollIntoViewOptions",
+    "ScrollBehavior",
+    "ScrollLogicalPosition",
+    "StorageManager",
+    "FileReader",
+    "EventInit",
+    "DataTransfer",
+    "File",
+    "FileList",
+    "ResizeObserver",
+    "ResizeObserverEntry",
+    "Performance",
+] }
+base64 = "0.22.0"
+leptos_router = {version = "0.6.11", features = ["csr"]}
+uiua = {path = "../..", default-features = false, features = ["batteries", "web"]}
+image = "0.24.9"
+unicode-segmentation = "1.10"
+urlencoding = "2"
+wasm-bindgen = "0.2.93"
+wasm-bindgen-futures = "0.4.43"
+hound = "3.5.1"
+
+# logging
+log = "0.4"
+wasm-logger = "0.2.0"
+console_error_panic_hook = "0.1.5"
