+++ site/Cargo.toml
@@ -19,6 +19,7 @@ pathdiff = "0.2.1"
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
 uiua = {path = "..", default-features = false, features = ["batteries", "web"]}
+uiua-editor = {path = "../pad/editor"}
 unicode-segmentation = "1.10"
 urlencoding = "2"
 wasm-bindgen = "0.2.93"
